Move local tracing out of the harness into src/tracing
The local trace spool (writer, reader, retention), the zero-config
local OTel runtime, and the agent span capture (provider, content
attributes, trace-state store, span-processor router) are a tracing
subsystem, not harness internals — the harness only feeds them
instrumentation events. They move to src/tracing, importable as
#tracing/*. The harness keeps the instrumentation vocabulary the tool
loop owns (lifecycle events, hooks, the AI SDK bridge, the runtime
registry); tracing depends on that surface and nothing depends back.
Guard rule 15's workflow-import ban now covers src/tracing alongside
channel and harness. No behavior change.
