+++ .changeset/tracing-module-move.md
@@ -1,0 +1,5 @@
+---
+"eve": patch
+---
+
+The local tracing subsystem (spool writer/reader, retention, the zero-config local OTel runtime, and agent span capture) moved from `src/harness` into its own `src/tracing` module; no behavior change.
