+++ packages/eve/src/cli/commands/trace.ts
@@ -2,12 +2,12 @@ import { basename } from "node:path";
 
 import { formatElapsed } from "#cli/format-elapsed.js";
 import { createCliTheme, renderCliSection, sanitizeForTerminal } from "#cli/ui/output.js";
-import type { LocalTrace, LocalTraceSpan } from "#harness/local-trace-reader.js";
+import type { LocalTrace, LocalTraceSpan } from "#tracing/local-trace-reader.js";
 import {
   compareLocalTraceSpans,
   describeLocalTraceSpan,
   listLocalTraces,
-} from "#harness/local-trace-reader.js";
+} from "#tracing/local-trace-reader.js";
 
 const TRACE_DISPLAY_DIRECTORY = ".eve/traces/v1";
