+++ packages/eve/src/cli/dev/tui/traces/trace-conversation.ts
@@ -14,8 +14,8 @@ import {
   visibleLength,
   wrapVisibleLine,
 } from "#cli/ui/terminal-text.js";
-import type { LocalTrace, LocalTraceSpan } from "#harness/local-trace-reader.js";
-import { compareLocalTraceSpans } from "#harness/local-trace-reader.js";
+import type { LocalTrace, LocalTraceSpan } from "#tracing/local-trace-reader.js";
+import { compareLocalTraceSpans } from "#tracing/local-trace-reader.js";
 
 import { formatCompactTokenCount } from "../stream-format.js";
 import type { Theme } from "../theme.js";
