+++ packages/eve/src/internal/nitro/host/local-tracing-runtime-plugin.ts
@@ -1,6 +1,6 @@
 import { basename } from "node:path";
 
-import { installLocalInstrumentationRuntime } from "#harness/local-instrumentation-runtime.js";
+import { installLocalInstrumentationRuntime } from "#tracing/local-instrumentation-runtime.js";
 import { resolveInstalledPackageInfo } from "#internal/application/package.js";
 import { DEVELOPMENT_WORKER_APP_ROOT_ENV } from "#internal/workflow/development-world-protocol.js";
