+++ packages/eve/src/tracing/agent-trace-context-store.ts
@@ -6,7 +6,7 @@ import type {
   AgentSessionTraceState,
   AgentTraceStateStore,
   AgentTurnTraceState,
-} from "#harness/agent-otel-provider.js";
+} from "#tracing/agent-otel-provider.js";
 
 interface AgentTraceContextState {
   readonly sessions: Readonly<Record<string, AgentSessionTraceState>>;
