+++ packages/eve/src/tracing/local-instrumentation-runtime.scenario.test.ts
@@ -8,10 +8,10 @@ import { afterEach, describe, expect, it } from "vitest";
 
 import { ContextContainer, contextStorage } from "#context/container.js";
 import { createAiSdkHookBridge } from "#harness/ai-sdk-hook-bridge.js";
-import { listLocalTraces } from "#harness/local-trace-reader.js";
+import { listLocalTraces } from "#tracing/local-trace-reader.js";
 import type { InstrumentationAttemptScope } from "#harness/instrumentation-lifecycle.js";
-import { installLocalInstrumentationRuntime } from "#harness/local-instrumentation-runtime.js";
-import { LocalTraceSpanProcessor } from "#harness/local-trace-span-processor.js";
+import { installLocalInstrumentationRuntime } from "#tracing/local-instrumentation-runtime.js";
+import { LocalTraceSpanProcessor } from "#tracing/local-trace-span-processor.js";
 
 const temporaryDirectories: string[] = [];
